+++ Cargo.toml
@@ -3,16 +3,24 @@ name = "webb-relayer"
 version = "0.4.0"
 authors = ["Shady Khalifa <shekohex@gmail.com>"]
 edition = "2021"
+description = "The Webb Relayer toolkit" 
+license = "Apache-2.0"
+
+
+[[bin]]
+name = "webb-relayer"
+path = "src/main.rs"
+required-features = ["cli"]
 
 [dependencies]
-anyhow = "^1"
+anyhow = { version = "^1", optional = true }
 parking_lot = "^0.12"
 once_cell = "^1.8"
 thiserror = "^1"
 async-trait = "^0.1"
 tracing = { version = "^0.1", features = ["log"] }
 tracing-futures = "^0.2"
-tracing-subscriber = { version = "0.3.1", features = ["parking_lot", "env-filter"] }
+tracing-subscriber = { version = "0.3.1", features = ["parking_lot", "env-filter"], optional = true }
 tracing-test = "0.2"
 url = { version = "^2.2", features = ["serde"] }
 sled = { version = "^0.34" }
@@ -25,9 +33,9 @@ warp = { version = "0.3.1", default-features = false, features = ["websocket", "
 warp-real-ip = "0.2"
 config = { version = "0.11", default-features = false, features = ["toml", "json"] }
 serde_json = { version = "^1", default-features = false }
-structopt = { version = "^0.3", features = ["paw"] }
-paw = "^1.0"
-directories-next = "^2.0"
+structopt = { version = "^0.3", features = ["paw"], optional = true }
+paw = { version = "^1.0", optional = true }
+directories-next = { version = "^2.0", optional = true }
 rand = { version = "0.8", default-features = false, features = ["getrandom"] }
 hex = { version = "0.4", default-features = false }
 # just to make it compile on linux where the openssl is not available.
@@ -53,6 +61,7 @@ tempfile = "^3.2"
 
 [features]
 default = ["evm-runtime", "substrate-runtime"]
+cli = ["evm-runtime", "substrate-runtime", "anyhow", "tracing-subscriber", "structopt", "paw", "directories-next"]
 evm-runtime = ["webb/evm-runtime", "webb-proposals/evm"]
 substrate-runtime = ["webb/substrate-runtime", "webb-proposals/substrate"]
 integration-tests = ["tracing-subscriber/json"]
