+++ src/types/mod.rs
@@ -1,4 +1,8 @@
+/// A custom type to support `Mnemonic` deserialization from environment variables.
 pub mod mnemonic;
+/// A custom type to support `PrivateKey` deserialization from environment variables.
 pub mod private_key;
+/// A custom type to support `RpcUrl` deserialization from environment variables.
 pub mod rpc_url;
+/// A custom type to support `Suri` deserialization from environment variables.
 pub mod suri;
